Prepare for adding MRP tests
This commit enables the MRP test suite. Additional (minimum) work has
been done to the fake MRP device to get connection establishment to
work, but no pure MRP tests are there yet. All common tests have been
moved to DMAP for now as that was the simplest solution. Once a test has
been implemented, it can be moved back to common tests again.
